+++ test/clj/rems/test_pdf.clj
@@ -146,11 +146,11 @@
                (fn []
                  (with-fixed-time (time/date-time 2010)
                    (fn []
-                     (#'pdf/render-application (applications/get-application handler application-id)))))))))
+                     (#'pdf/render-application (applications/get-application-for-user handler application-id)))))))))
       (testing "pdf rendering succeeds"
         (is (some?
              (with-language :en
                #(do
                   ;; uncomment this to get a pdf file to look at
-                  #_(pdf/application-to-pdf (applications/get-application handler application-id) "/tmp/example-application.pdf")
-                  (pdf/application-to-pdf-bytes (applications/get-application handler application-id)))))))))
+                  #_(pdf/application-to-pdf (applications/get-application-for-user handler application-id) "/tmp/example-application.pdf")
+                  (pdf/application-to-pdf-bytes (applications/get-application-for-user handler application-id)))))))))